Move BaseShareSerializer to api/serializer.py
Check if proper route
