+++ api/urls.py
@@ -1,18 +1,15 @@
-from collections import OrderedDict
-
 from django.conf.urls import url
 from django.views.decorators.csrf import csrf_exempt
 from django.views.decorators.csrf import ensure_csrf_cookie
 
 from rest_framework.routers import DefaultRouter
-from rest_framework_json_api import serializers
 
 from graphene_django.views import GraphQLView
 
 from share import models
 
 from api import views
-from api import fields
+from api.serializers import BaseShareSerializer
 from api.views.share import ShareObjectViewSet
 
 router = DefaultRouter()
@@ -20,52 +17,6 @@ router = DefaultRouter()
 
 def register_route(prefix, viewset):
     router.register(prefix, viewset, base_name=viewset.serializer_class.Meta.model._meta.model_name)
-
-
-class BaseShareSerializer(serializers.ModelSerializer):
-
-    def __init__(self, *args, **kwargs):
-        # super hates my additional kwargs
-        sparse = kwargs.pop('sparse', False)
-        version_serializer = kwargs.pop('version_serializer', False)
-        super(BaseShareSerializer, self).__init__(*args, **kwargs)
-
-        if sparse:
-            # clear the fields if they asked for sparse
-            self.fields.clear()
-        else:
-            # remove hidden fields
-            excluded_fields = ['change', 'uuid', 'sources']
-            for field_name in tuple(self.fields.keys()):
-                if 'version' in field_name or field_name in excluded_fields:
-                    self.fields.pop(field_name)
-
-            if not version_serializer:
-                # add links to related objects
-                self.fields.update({
-                    'links': fields.LinksField(links=self.Meta.links, source='*')
-                })
-
-        # version specific fields
-        if version_serializer:
-            self.fields.update({
-                'action': serializers.CharField(max_length=10),
-                'persistent_id': serializers.IntegerField()
-            })
-
-        # add fields with improper names
-        self.fields.update({
-            'type': fields.TypeField(),
-        })
-
-    class Meta:
-        links = ('versions', 'changes', 'rawdata')
-
-    # http://stackoverflow.com/questions/27015931/remove-null-fields-from-django-rest-framework-response
-    def to_representation(self, instance):
-        ret = super(BaseShareSerializer, self).to_representation(instance)
-        ret = OrderedDict(list(filter(lambda x: x[1] is not None, ret.items())))
-        return ret
 
 
 class EndpointGenerator:
